refactor: moved sftp logic to seperate module
