+++ server/meca/index.js
@@ -1,5 +1,3 @@
-const SftpClient = require('ssh2-sftp-client')
-const config = require('config')
 const ManuscriptManager = require('@elifesciences/xpub-server/entities/manuscript')
 
 const articleGenerator = require('./file-generators/article')
@@ -9,6 +7,8 @@ const manifestGenerator = require('./file-generators/manifest')
 const manuscriptGenerator = require('./file-generators/manuscript')
 const transferGenerator = require('./file-generators/transfer')
 const archiveGenerator = require('./file-generators/archive')
+
+const upload = require('./upload')
 
 async function generate(manuscriptId, userId, clientIp) {
   const manuscript = await ManuscriptManager.find(manuscriptId, userId)
@@ -23,20 +23,9 @@ async function generate(manuscriptId, userId, clientIp) {
   })
 }
 
-async function deliver(file, id) {
-  const sftp = new SftpClient()
-  await sftp.connect(config.get('meca.sftp'))
-
-  const remotePath = config.get('meca.remotePath')
-  await sftp.mkdir(remotePath, true)
-
-  await sftp.put(file, `${remotePath}/${id}`)
-  await sftp.end()
-}
-
-async function mecaExport(manuscriptId, userId, clientIp) {
+async function send(manuscriptId, userId, clientIp) {
   const archive = await generate(manuscriptId, userId, clientIp)
-  await deliver(archive, manuscriptId)
+  await upload(archive, manuscriptId)
 }
 
 module.exports = mecaExport
