+++ server/meca/upload/index.js
@@ -1,0 +1,7 @@
+const { uploadToSFTP } = require('./sftp');
+
+async function upload(file, id) {
+  await uploadToSFTP(file, id)
+}
+
+module.exports = upload
